Piece square tables

+++ JenjinnV2/src/jenjinn/engine/evaluation/BoardEvaluator.java
@@ -9,6 +9,10 @@ package jenjinn.engine.evaluation;
 import jenjinn.engine.boardstate.BoardState;
 
 /**
+ * Simple interface representing an object which can take a {@link BoardState} instance
+ * as an argument and calculate an unsigned score evaluation. By unsigned we mean that
+ * no matter who it is to move, a higher score is better for them.
+ *
  * @author ThomasB
  * @since 27 Jul 2017
  */
@@ -16,14 +20,14 @@ public interface BoardEvaluator
 {
 	/**
 	 * A SYMMETRIC evaluation function taking a {@link BoardState} instance as a
-	 * parameter and returning a 32bit integer value where the higher the score
+	 * parameter and returning a 16bit integer value where the higher the score
 	 * the better for the side which has the move and the lower the score the
 	 * better for the opposition.
 	 *
 	 * @param state is what we want to evaluate.
 	 * @return the evaluation value
 	 */
-	int evaluate(BoardState state);
+	short evaluate(BoardState state);
 }
 
 /* ---------------------------------------------------------------------*
